+++ _codegen/cog_sources/ChunksmithFabric.java
@@ -164,6 +164,9 @@ public class ChunksmithFabric implements ModInitializer {
                 argument(CommandLiteral.RADIUS, word()));
         registerArguments(command, literal(CommandLiteral.RELOAD),
                 argument(CommandLiteral.TYPE, word()));
+        registerArguments(command, literal(CommandLiteral.SET),
+                argument(CommandLiteral.TYPE, string()).suggests(SuggestionProviders.SETTINGS),
+                argument(CommandLiteral.VALUE, string()));
         registerArguments(command, literal(CommandLiteral.SELECTION));
         registerArguments(command, literal(CommandLiteral.SHAPE),
                 argument(CommandLiteral.SHAPE, string()).suggests(SuggestionProviders.SHAPES));