+++ shared_common/src/main/java/com/kishku7/chunksmith/command/HelpCommand.java
@@ -27,6 +27,7 @@ public class HelpCommand implements ChunksmithCommand {
             CommandLiteral.PATTERN,
             CommandLiteral.SILENT,
             CommandLiteral.QUIET,
+            CommandLiteral.SET,
             CommandLiteral.TRIM,
             CommandLiteral.SELECTION,
             CommandLiteral.PROGRESS,